requirements.txt
streamlit==1.29.0
openai==1.3.0
python-docx==0.8.11
pyannote.audio==3.1.1
torch>=2.0.0
torchaudio>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
scipy>=1.7.0
huggingface_hub>=0.12.0
pytorch-metric-learning>=2.0.0
matplotlib>=3.5.0
librosa>=0.9.0
soundfile>=0.10.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
httpx==0.27.2 